Image/pyproject.toml
[project]
name = "image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "fhidargue", email = "fi.ha@hotmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.4",
    "pillow>=12.0.0",
]

[project.scripts]
image = "image:main"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
